% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCFA.R
\docType{data}
\name{GBM}
\alias{GBM}
\title{GBM}
\format{A list with two items:
\describe{
  \item{data}{List of microRNA data matrix.}
  \item{survival}{Survival information.}
}}
\usage{
GBM
}
\description{
GBM dataset, including microRNA and survidal data.
}
\keyword{datasets}
