% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SCFA.R
\name{plot.SCFA}
\alias{plot.SCFA}
\title{Hierarchical clustering of variables from Signal Cancellation Factor Analysis}
\usage{
\method{plot}{SCFA}(x, alpha = c(0.01, 0.05, 0.2), ...)
}
\arguments{
\item{x}{An object of class SCFA.}

\item{alpha}{Type I error rate to display for clustering of variables. Default is \eqn{\alpha = .01,.05,.20}.}

\item{...}{Further arguments for other methods, ignored for "scfa".}
}
\value{
A plot output.
}
\description{
Draw the hierarchical clustering of variables from Signal Cancelling Factor Analysis.
}
\examples{
plot(SCFA(R = Achim24, N = 1000))

}
