% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SCFA.R
\name{print.SCFA}
\alias{print.SCFA}
\title{Print results of Signal Cancellation Factor Analysis}
\usage{
\method{print}{SCFA}(x, ...)
}
\arguments{
\item{x}{An object of class "SCFA".}

\item{...}{Further arguments for other methods, ignored for "SCFA".}
}
\value{
No return value, called for side effects.
}
\description{
Print the number of factors to retain.
}
\examples{
results <- SCFA(TabachnickFidell, N = 175)
print(results)
}
