% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCFA.R
\name{SCFA}
\alias{SCFA}
\title{SCFA}
\usage{
SCFA(dataList, k = NULL, max.k = 5, ncores = 10L, seed = NULL)
}
\arguments{
\item{dataList}{List of data matrices. In each matrix, rows represent samples and columns represent genes/features.}

\item{k}{Number of clusters, leave as default for auto detection.}

\item{max.k}{Maximum number of cluster}

\item{ncores}{Number of processor cores to use.}

\item{seed}{Seed for reproducibility.}
}
\value{
A numeric vector containing cluster assignment for each sample.
}
\description{
The main function to perform subtyping
}
