% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SCFA.R
\name{summary.SCFA}
\alias{summary.SCFA}
\title{Summary of Signal Cancellation Factor Analysis}
\usage{
\method{summary}{SCFA}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{An object of class "SCFA".}

\item{alpha}{Type I error rate to display meaningful loadings and correlations between factors.}

\item{...}{Further arguments for other methods, ignored for "SCFA".}
}
\value{
No return value, called for side effects.
}
\description{
Summary of Signal Cancellation Factor Analysis
}
\examples{
summary(SCFA(R = Achim24, N = 1000))
}
