% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCFA.R
\name{SCFA.class}
\alias{SCFA.class}
\title{SCFA.class}
\usage{
SCFA.class(dataListTrain, trainLabel, dataListTest, ncores = 10L,
  seed = NULL)
}
\arguments{
\item{dataListTrain}{List of training data matrices. In each matrix, rows represent samples and columns represent genes/features.}

\item{trainLabel}{Survival information of patient in training set in form of Surv object.}

\item{dataListTest}{List of testing data matrices. In each matrix, rows represent samples and columns represent genes/features.}

\item{ncores}{Number of processor cores to use.}

\item{seed}{Seed for reproducibility.}
}
\value{
A vector of risk score predictions for patient in test set.
}
\description{
This function provides low dimensional representations for risk prediction.
}
