% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCFA.R
\name{SCFA}
\alias{SCFA}
\title{Signal Cancellation Factor Analysis}
\usage{
SCFA(R, N = NULL, alpha = 0.05, factors.only = FALSE)
}
\arguments{
\item{R}{A correlation matrix or a data set.}

\item{N}{The sample size if a correlation matrix is supplied.}

\item{alpha}{Type I error rate on the hierarchical clustering of variables.}

\item{factors.only}{An option if only the number of dimensions is required.}
}
\value{
An object of class SCFA.
}
\description{
Signal Cancellation Factor Analysis
}
\examples{
SCFA(R = Achim24, N = 1000)

}
\references{
Achim, A. (2024, April 4). Signal cancellation factor analysis. \emph{PsyArXiv}, 1–13. \doi{10.31234/osf.io/h7qwg}
}
